+++ src/main/java/com/myhome/repository/LetterToADMINRepository.java
@@ -5,6 +5,6 @@ import com.myhome.models.LetterToADMIN;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface LetterToADMINRepository extends JpaRepository<LetterToADMIN, Long> {
+public interface LetterToADMINRepository extends JpaRepository<LetterToADMIN, Integer> {
 
 }
